% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robocopy.R
\name{sfs_robocopy_available}
\alias{sfs_robocopy_available}
\title{Check whether \code{robocopy} is available}
\usage{
sfs_robocopy_available()
}
\value{
A single logical value.
}
\description{
Checks whether \code{robocopy} is on \code{PATH}. Unlike
\code{\link[=sfs_powershell_available]{sfs_powershell_available()}}, does not test that it can be run.
}
\examples{
sfs_robocopy_available()
}
