% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{sfs_stage_cleanup}
\alias{sfs_stage_cleanup}
\title{Remove a staging directory created by \code{sfs_stage_local()}}
\usage{
sfs_stage_cleanup(staged)
}
\arguments{
\item{staged}{The tibble returned by \code{\link[=sfs_stage_local]{sfs_stage_local()}}.}
}
\value{
\code{TRUE} (invisibly).
}
\description{
Remove a staging directory created by \code{sfs_stage_local()}
}
