% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharefs-package.R
\docType{package}
\name{sharefs-package}
\alias{sharefs}
\alias{sharefs-package}
\title{sharefs: Fast Filesystem Operations for Network Shares}
\description{
Alternatives to base R's file listing, metadata, and copy functions, optimized specifically for reading files off an SMB/CIFS network share (a mapped network drive or UNC path) on Windows, where each individual file operation -- listing a directory, calling 'stat()' on a file, copying a file -- pays a full network round trip. 'sharefs' uses a single 'Get-ChildItem()' call via 'PowerShell' to list a directory together with file metadata, and 'robocopy' to stage files locally, instead of one call per file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/orgadish/sharefs}
  \item Report bugs at \url{https://github.com/orgadish/sharefs/issues}
}

}
\author{
\strong{Maintainer}: Or Gadish \email{orgadish@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Or Gadish \email{orgadish@gmail.com} [copyright holder]
}

}
\keyword{internal}
