% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir-info.R
\name{sfs_dir_info}
\alias{sfs_dir_info}
\title{List files with minimal network calls}
\usage{
sfs_dir_info(
  path = ".",
  all = FALSE,
  recurse = FALSE,
  type = "any",
  regexp = NULL,
  glob = NULL,
  invert = FALSE,
  fail = TRUE,
  ...
)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{all}{If \code{TRUE} hidden files are also returned.}

\item{recurse}{Whether to recurse into subdirectories. Depth-limited
recursion is not supported.}

\item{type}{File type(s) to return, one or more of "any", "file",
"directory", "symlink", "FIFO", "socket", "character_device" or
"block_device".}

\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to
\code{\link[=grep]{grep()}} to filter paths.}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on
to \code{\link[=grep]{grep()}} to filter paths.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{fail}{Should the call fail (the default) or warn if a file
cannot be accessed.}

\item{...}{Passed to \code{\link[=grepl]{grepl()}} (e.g. \code{ignore.case = TRUE}).}
}
\value{
A data.frame with columns \code{path}, \code{type}, \code{size},
\code{modification_time}, \code{access_time}, \code{birth_time} (or \code{tibble}
if installed).
}
\description{
Lists files and metadata in a single network round trip, using
PowerShell's \code{Get-ChildItem} instead of one \code{stat()} per file.

This is similar to \code{\link[fs:dir_info]{fs::dir_info()}}, but not identical.

The output includes only the subset of the columns that are available via
\code{Get-ChildItem}: \code{path}, \code{type}, \code{size}, \code{modification_time}, \code{access_time},
\code{birth_time}. Unlike \code{\link[fs:dir_info]{fs::dir_info()}}, it will return partial results with
a \code{sharefs_warning_powershell_partial} warning if some items had to be
skipped (e.g. a permission-denied subfolder).

Automatically includes retry logic if an intermittent error is encountered.
}
\examples{
if (sfs_powershell_available()) {
  sfs_dir_info(system.file(package = "sharefs"))
}
}
\seealso{
\code{\link[=sfs_powershell_available]{sfs_powershell_available()}} which checks if Powershell is
available and accessible.
}
