% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir-info.R
\name{sfs_dir_ls}
\alias{sfs_dir_ls}
\title{List file paths with minimal network calls}
\usage{
sfs_dir_ls(
  path = ".",
  all = FALSE,
  recurse = FALSE,
  type = "any",
  regexp = NULL,
  glob = NULL,
  invert = FALSE,
  fail = TRUE,
  ...
)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{all}{If \code{TRUE} hidden files are also returned.}

\item{recurse}{Whether to recurse into subdirectories. Depth-limited
recursion is not supported.}

\item{type}{File type(s) to return, one or more of "any", "file",
"directory", "symlink", "FIFO", "socket", "character_device" or
"block_device".}

\item{regexp}{A regular expression (e.g. \verb{[.]csv$}) passed on to
\code{\link[=grep]{grep()}} to filter paths.}

\item{glob}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on
to \code{\link[=grep]{grep()}} to filter paths.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{fail}{Should the call fail (the default) or warn if a file
cannot be accessed.}

\item{...}{Passed to \code{\link[=grepl]{grepl()}} (e.g. \code{ignore.case = TRUE}).}
}
\value{
A named \code{fs::fs_path} character vector.
}
\description{
\code{sfs_dir_info(...)$path}. No extra cost to fetching full metadata even
when only paths are needed, since \code{Get-ChildItem} returns it either
way.
}
\examples{
if (sfs_powershell_available()) {
  sfs_dir_ls(system.file(package = "sharefs"))
}
}
